+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/stretching/AutohistogramStrategy.java
@@ -19,6 +19,7 @@ import me.champeau.a4j.jsolex.processing.sun.BackgroundRemoval;
 import me.champeau.a4j.jsolex.processing.sun.Broadcaster;
 import me.champeau.a4j.jsolex.processing.sun.tasks.ImageAnalysis;
 import me.champeau.a4j.jsolex.processing.sun.workflow.AnalysisUtils;
+import me.champeau.a4j.jsolex.processing.util.Constants;
 import me.champeau.a4j.jsolex.processing.util.Histogram;
 import me.champeau.a4j.jsolex.processing.util.ImageWrapper32;
 import me.champeau.a4j.math.regression.Ellipse;
@@ -26,16 +27,16 @@ import me.champeau.a4j.math.regression.Ellipse;
 import java.util.ArrayList;
 import java.util.List;
 
-import static me.champeau.a4j.jsolex.processing.util.Constants.MAX_PIXEL_VALUE;
 import static me.champeau.a4j.jsolex.processing.util.Constants.message;
 import static me.champeau.a4j.jsolex.processing.util.LoggingSupport.LOGGER;
 
 public final class AutohistogramStrategy implements StretchingStrategy {
     private static final int HISTOGRAM_BINS = 256;
-    private static final int FULL_RESOLUTION_BINS = 65536;
+    private static final double TARGET_PEAK = 0.65;
+    private static final double LIMB_MARGIN = 1.1;
+    private static final double REDUCED_AMPLIFICATION_LIMIT = 1.3;
 
     public static final double DEFAULT_GAMMA = 1.5;
-    public static final double TARGET_PEAK = 0.65;
 
     private final double gamma;
 
@@ -66,40 +67,53 @@ public final class AutohistogramStrategy implements StretchingStrategy {
         var width = image.width();
         if (ellipse.isPresent()) {
             var e = ellipse.get();
+            disk = BackgroundRemoval.neutralizeBackground(disk, 4);
+            diskData = disk.data();
             // the initial strech is to adjust the global brightness by estimating the brightness of
             // the disk itself
-            var avg = 0d;
-            var area = 0d;
             var stats = ImageAnalysis.masked(diskData, width, height, e);
+            var cx = e.center().a();
+            var cy = e.center().b();
+            var semiAxis = e.semiAxis();
+            var radius = (semiAxis.a() + semiAxis.b()) / 2;
+            var amplificationThreshold = 0.5 * AnalysisUtils.estimateBlackPoint(disk, e);
             if (stats.avg() > stats.max() / 8) {
-                new GammaStrategy(gamma).stretch(disk);
-                var background = AnalysisUtils.estimateBlackPoint(disk, e) * 1.2f;
+                float max = 1e-7f;
+                for (float v : diskData) {
+                    max = Math.max(v, max);
+                }
                 for (int y = 0; y < height; y++) {
                     for (int x = 0; x < width; x++) {
-                        var v = diskData[y * width + x];
-                        if (e.isWithin(x, y) && v > background) {
-                            avg += v;
-                            area++;
+                        var idx = y * width + x;
+                        var v = diskData[idx];
+                        float normalized = v / max;
+                        var dist = normalizedDistanceToCenter(x, y, cx, cy, radius);
+                        var gammaCorrected = (float) Math.pow(normalized, gamma) * Constants.MAX_PIXEL_VALUE;
+                        if (dist <= 1 || v < amplificationThreshold) {
+                            diskData[idx] = gammaCorrected;
+                        } else {
+                            float corrected = (float) Math.pow((gammaCorrected - amplificationThreshold) / max, Math.max(.5, gamma * (.2 * Math.pow(1 / dist, .5))));
+                            float newValue = (corrected * Constants.MAX_PIXEL_VALUE);
+                            if (Float.isNaN(newValue)) {
+                                newValue = gammaCorrected;
+                            }
+                            var diff = newValue - gammaCorrected;
+                            float limitFactor = 1f;
+                            if (dist < LIMB_MARGIN) {
+                                var k = (LIMB_MARGIN - dist) / (LIMB_MARGIN - 1);
+                                limitFactor = (float) (1f - k);
+                            }
+                            if (dist > REDUCED_AMPLIFICATION_LIMIT) {
+                                limitFactor /= (float) (4 * (1 + (dist - REDUCED_AMPLIFICATION_LIMIT)));
+                            }
+                            newValue = gammaCorrected + diff * limitFactor;
+                            diskData[idx] = newValue;
                         }
                     }
-                }
-                avg /= area;
-                // compute stdev
-                var stdev = 0d;
-                for (int y = 0; y < height; y++) {
-                    for (int x = 0; x < width; x++) {
-                        var v = diskData[y * width + x];
-                        if (e.isWithin(x, y) && v > background) {
-                            stdev += (v - avg) * (v - avg);
-                        }
-                    }
-                }
-                stdev = Math.sqrt(stdev / area);
-                var limit = 1.2 * (avg + 6 * stdev);
-                if (limit < MAX_PIXEL_VALUE) {
-                    new ContrastAdjustmentStrategy(0, (float) limit).stretch(disk);
                 }
             } else {
+                disk = image.copy();
+                diskData = disk.data();
                 for (int i = 0; i < diskData.length; i++) {
                     float v = diskData[i];
                     diskData[i] = (float) Math.pow(v, 0.8);
@@ -108,20 +122,10 @@ public final class AutohistogramStrategy implements StretchingStrategy {
                 LOGGER.warn(message("skip.gamma.stretch"));
             }
 
-            var mask = createMask(height, width, e);
-            var protus = prepareProtusImage(image, e);
-            var protusData = protus.data();
-            for (int y = 0; y < height; y++) {
-                for (int x = 0; x < width; x++) {
-                    var idx = y * width + x;
-                    var weight = mask[idx];
-                    diskData[idx] = ((1 - weight) * disk.data()[idx] + weight * protusData[idx]);
-                }
-            }
-            System.arraycopy(disk.data(), 0, image.data(), 0, diskData.length);
+            System.arraycopy(diskData, 0, image.data(), 0, diskData.length);
         } else {
             new GammaStrategy(gamma).stretch(disk);
-            System.arraycopy(disk.data(), 0, image.data(), 0, diskData.length);
+            System.arraycopy(diskData, 0, image.data(), 0, diskData.length);
         }
         diskData = image.data();
         // for histogram transform, we will only consider pixels within 1.2 * radius of the center of the disk
@@ -142,82 +146,15 @@ public final class AutohistogramStrategy implements StretchingStrategy {
         }
     }
 
-    private static ImageWrapper32 prepareProtusImage(ImageWrapper32 image, Ellipse ellipse) {
-        var protus = image.copy();
-        var data = protus.data();
-        trimLeft(data);
-        protus = BackgroundRemoval.neutralizeBackground(protus);
-        var blackPoint = AnalysisUtils.estimateBlackPoint(protus, ellipse);
-        new StretchingChain(
-            new GammaStrategy(0.5),
-            new ArcsinhStretchingStrategy((float) blackPoint, 5, 5)
-        ).stretch(protus);
-        // Truncate the histogram on the left to protect shadows
-        var histo = Histogram.of(data, 256);
-        for (int i = 0; i < 32; i++) {
-            if (histo.get(i) == 0) {
-                for (int j = i + 1; j < 32; j++) {
-                    if (histo.get(j) == 0) {
-                        i++;
-                    }
-                }
-                truncateLeft(data, 256f * i);
-                break;
-            }
-        }
-        return protus;
-    }
-
-    private static void trimLeft(float[] data) {
-        float minNonZero = Float.MAX_VALUE;
-        for (float v : data) {
-            if (v > 0) {
-                minNonZero = Math.min(v, minNonZero);
-            }
-        }
-        // normalize so that the minimum is non-zero (which can happen because of artifacts or resizing, ...)
-        truncateLeft(data, minNonZero);
-    }
-
-    private static void truncateLeft(float[] data, float minNonZero) {
-        for (int i = 0; i < data.length; i++) {
-            data[i] = Math.max(0, data[i] - minNonZero);
-        }
-    }
-
     // computes the distance to the circle center, relative to the radius. A negative value
     // means that the point is inside the circle, a positive value means that the point is outside
     // the circle. The distance is normalized so that it is 0 at the circle border, and 1 at the
     // circle center.
     private static double normalizedDistanceToCenter(double x, double y, double cx, double cy, double radius) {
-        double distance = Math.sqrt((x - cx) * (x - cx) + (y - cy) * (y - cy));
-        return (distance - radius) / radius;
-    }
-
-    private float[] createMask(int height, int width, Ellipse e) {
-        float[] mask = new float[width * height];
-        var cx = e.center().a();
-        var cy = e.center().b();
-        var semiAxis = e.semiAxis();
-        var radius = (semiAxis.a() + semiAxis.b()) / 2;
-        for (int y = 0; y < height; y++) {
-            for (int x = 0; x < width; x++) {
-                var idx = y * width + x;
-                var distance = normalizedDistanceToCenter(x, y, cx, cy, radius);
-                var v = 0.5 * (distance < 0.005 ? Math.pow(Math.exp(distance - 0.01), 24) : 1 - distance);
-                mask[idx] = (float) (v * v);
-            }
-        }
-
-//        if (broadcaster != null) {
-//            var t = new float[width * height];
-//            System.arraycopy(mask, 0, t, 0, mask.length);
-//            for (int i = 0; i < t.length; i++) {
-//                t[i] = t[i] * MAX_PIXEL_VALUE;
-//            }
-//            broadcaster.broadcast(new ImageGeneratedEvent(new GeneratedImage(GeneratedImageKind.IMAGE_MATH, "Mask", Path.of("/tmp/mask.png"), new ImageWrapper32(width, height, t, Map.of()))));
-//        }
-        return mask;
+        var dx = x - cx;
+        var dy = y - cy;
+        double distance = Math.sqrt(dx * dx + dy * dy);
+        return distance / radius;
     }
 
     private static Histogram maskedHistogram(float[] image, int width, int height, Ellipse ellipse) {